requirements.txt
APScheduler==3.6.3
attrs==19.3.0
autopep8==2.0.2
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2019.9.11
chardet==3.0.4
contourpy==1.0.7
cycler==0.11.0
Django==2.2.6
django-debug-toolbar==2.2
Faker==12.0.1
fonttools==4.39.3
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
mixer==7.1.2
more-itertools==8.2.0
numpy==1.24.3
packaging==20.1
Pillow==9.5.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.10.0
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2019.3
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.4.1
sqlparse==0.3.0
tomli==2.0.1
tornado==6.3.2
tzlocal==5.0.1
urllib3==1.25.6
wcwidth==0.1.8
zipp==3.15.0
python-dotenv==1.0.0